Question_4/requirements.txt
Flask==3.0.3
numpy==1.24.3
pytest==8.3.2
h5py==3.11.0
werkzeug==3.0.3
Pillow==10.4.0